refactor(types): 将边值类型定义移动到共享类型文件
将ShorthandMode枚举和EdgeValues接口从组件背景类型文件移动到核心类型文件，以便其他模块复用
